+++ src/main/java/com/ncasanovas/crud/repository/RelacionRepository.java
@@ -3,5 +3,5 @@ package com.ncasanovas.crud.repository;
 import com.ncasanovas.crud.entity.RelacionEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RelacionRepository  extends CrudRepository<RelacionEntity, Long> {
+public interface RelacionRepository  extends CrudRepository<RelacionEntity, Integer> {
 }
